+++ components/InvoiceEdit/InvoiceEditBuyer.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { useRecoilState } from 'recoil';
-import BuyerInput from '../components/BuyerInput';
+import BuyerInput from '../inputs/BuyerInput';
 
-import { buyerState, emailState, lockedState } from '../src/atoms';
+import { buyerState, emailState, lockedState } from '../../src/atoms';
 
 export default function InvoiceEditBuyer() {
   const [buyer, setBuyer] = useRecoilState(buyerState);
