+++ components/InvoiceEdit/IssuerInput.tsx
@@ -3,12 +3,12 @@ import TextField from '@material-ui/core/TextField';
 import { useRecoilState } from 'recoil';
 import useSWR from 'swr';
 
-import { cleanUpString } from '../utils/textutils';
+import { cleanUpString } from '../../utils/textutils';
 import {
   languageAfterChangeState,
   lockedState,
   issuerState,
-} from '../src/atoms';
+} from '../../src/atoms';
 
 export default function IssuerInput() {
   const [issuer, setIssuer] = useRecoilState(issuerState);
