+++ components/InvoiceEdit/SendInvoiceButton.tsx
@@ -4,7 +4,7 @@ import Button from '@material-ui/core/Button';
 import Grid from '@material-ui/core/Grid';
 import Typography from '@material-ui/core/Typography';
 
-import Link from '../src/Link';
+import Link from '../../src/Link';
 import {
   emailState,
   invoiceIdState,
@@ -12,7 +12,7 @@ import {
   messageSeverityState,
   messageTextState,
   sentState,
-} from '../src/atoms';
+} from '../../src/atoms';
 
 interface IProps {
   disabled: boolean;
