+++ components/InvoiceEdit/SeriesIdInput.tsx
@@ -10,7 +10,7 @@ import {
   lockedState,
   seriesIdState,
   seriesNameState,
-} from '../src/atoms';
+} from '../../src/atoms';
 
 export default function SeriesIdInput() {
   const [initialInvoice] = useRecoilState(initialInvoiceState);
